Stow an RDMA access mode in each buflist and use that mode to register each buffer.
Flesh out my sink implementation, making it verify that the content in a
buffer FIFO matches its expectations.

Add a FIFO to `rcvr_t`, `rxfifo`, that tracks which buffers await RDMA
writes from the peer.

Make both peers fi_shutdown() their connection and wait for an
FI_SHUTDOWN event before exiting.  Move some event queues from the
get_/put_state_t to the rcvr_t/xmtr_t so that they're available in the
receive/transmit loops.

Write a helper routine, write_fully(), that performs an RDMA write in
as many parts as the RDMA I/O vector list limit requires.  Use it in
xmtr_loop().

Replace a bunch of printf(3) with warnx(3) so that messages are labeled
with the program that wrote them and they all go to stderr.

Initialize a `buflist_t` with a longer sequence of shorter buffers.

Assign sessions (connection/terminal pairs) to workers instead of
connections.