Cope with fi_poll(3) being unavailable as it is when we use the `cxi` provider.
Request tagged-message capability.

Bug fix: check `fi_domain()` return.

Cope with the `cxi` provider's requirement that memory is registered
on a per-endpoint basis.  This required some refactoring because some
memory was registered before its endpoint was created and I want for all
steps of registration to occur at about the same time so that no memory
is in an almost-but-not-quite-registered state.

Cope with the `cxi` provider's inability to deal with some/all (?)
`fi_trecvmsg` `ignore` masks: use 0, always.

Remove the `-b` option from `fget` and don't try to connect by
node/service name.  Instead, make `fget` always bind whichever address
the provider prefers, print that address as a hex string, and then let
`fput` convert a hex string on its single command-line argument to a
binary address and try to connect to that.  This works fine with both
the `tcp` and `cxi` providers, but I will have to make some changes to
`scripts/run-suite` so that it will work.

Diagnostics: label each buffer (bufhdr_t) with the endpoint where it was
registered just in case that's helpful in the debugger---hasn't been, so
far.  Add diagnostic logging statements here and there.  Log a bunch of
libfabric parameters (fi_info and such).  Provide `xfc_type_to_string`
for turning transfer-context types to strings and use it here and there.
Correct some error/log messages.

With the `cxi` provider, `fi_close` gripes about resources in-use when
we close an endpoint.  Add `session_shutdown()` and use it to clean up
some resources that may be bound to an endpoint, first, then close it.
Also, cancel all outstanding I/O before calling `session_shutdown`.

Sometimes `fi_av_insert` will fail by returning `rc == 0` instead of
returning `rc < 0`.  The `cxi` provider seems to do that.  Check for
that.